Removes uneccesary class
