+++ src/main/java/httpserver/Main.java
@@ -2,7 +2,6 @@ package httpserver;
 
 import httpserver.httpmessages.HTTPRequestHandler;
 import httpserver.httpmessages.HTTPRequestParser;
-import httpserver.httpmessages.HTTPResponseBuilder;
 import httpserver.resourcemanagement.HTTPResourceHandler;
 import httpserver.resourcemanagement.ResourceParser;
 import httpserver.routing.*;
@@ -26,7 +25,7 @@ public class Main {
                 String resources = "/Users/molliestephenson/Java/Server/HTTPServer/src/main/resources/form";
                 HTTPResourceHandler resourceHandler = new HTTPResourceHandler(path, new ResourceParser());
                 List<Route> registeredRoutes = Arrays.asList(new CoffeeRoute("/coffee", GET), new TeaRoute("/tea", GET), new MethodOptionsRoute("/method_options", GET, POST, PUT, OPTIONS, HEAD), new FormRoute("/form", resources, GET, POST, PUT, DELETE));
-                HTTPRequestHandler httpRequestHandler = new HTTPRequestHandler(new HTTPResponseBuilder(resourceHandler), new Router(registeredRoutes));
+                HTTPRequestHandler httpRequestHandler = new HTTPRequestHandler(resourceHandler, new Router(registeredRoutes));
                 HTTPRequestParser httpRequestParser = new HTTPRequestParser();
                 httpServer = new HTTPServer(socketServer, httpRequestHandler, httpRequestParser);
             } catch (IOException e) {
